+++ .eslintrc.js
@@ -27,7 +27,6 @@ module.exports = {
     '@typescript-eslint/no-var-requires': 'off',
 
     'import/no-unresolved': 'off',
-    'import/no-default-export': 2,
     'import/no-named-as-default-member': 'off',
     'import/export': 'off'