+++ examples/my-app/metro.config.js
@@ -1,0 +1,37 @@
+// Learn more https://docs.expo.dev/guides/monorepos
+const { getDefaultConfig } = require('expo/metro-config');
+const path = require('path');
+
+// Find the project and workspace directories
+const projectRoot = __dirname;
+
+// This can be replaced with `find-yarn-workspace-root`
+const workspaceRoot = path.resolve(projectRoot, '../..');
+
+// Only list the packages within your monorepo that your app uses. No need to add anything else.
+// If your monorepo tooling can give you the list of monorepo workspaces linked
+// in your app workspace, you can automate this list instead of hardcoding them.
+const monorepoPackages = {
+    '@0xsodium/provider': path.resolve(workspaceRoot, 'packages/provider'),
+};
+
+const config = getDefaultConfig(projectRoot);
+
+// 1. Watch all files within the monorepo
+config.watchFolders = [workspaceRoot];
+// 2. Let Metro know where to resolve packages and in what order
+config.resolver.nodeModulesPaths = [
+    path.resolve(projectRoot, 'node_modules'),
+    path.resolve(workspaceRoot, 'node_modules'),
+    path.resolve('/Users/alberthuang/code/graphql-mesh/node_modules'),
+];
+config.resolver.resolverMainFields = [
+    "react-native",
+    "browser",
+    "main"
+];
+config.resolver.extraNodeModules = monorepoPackages;
+// 3. Force Metro to resolve (sub)dependencies only from the `nodeModulesPaths`
+config.resolver.disableHierarchicalLookup = true;
+
+module.exports = config;